Update checkpointing logic
The old version of the logic for when to save a checkpoint was pretty
buggy and didn't respect the count of checkpoints to save. This version
calculates the step numbers to save at in advance so we know the count
is exactly right and they're as close as possible to evenly spaced
within the training process.